[CCUBE-2379] move the validation into numeric-field
